Fix up noisy compile warnings

+++ Bit Slicer/ZGMemoryDumpAllWindowController.m
@@ -142,12 +142,12 @@
 - (void)progressWillBegin:(ZGSearchProgress *)searchProgress
 {
 	_searchProgress = searchProgress;
-	_progressIndicator.maxValue = _searchProgress.maxProgress;
+	_progressIndicator.maxValue = (double)_searchProgress.maxProgress;
 }
 
 - (void)progress:(ZGSearchProgress *)searchProgress advancedWithResultSets:(NSArray<NSData *> *)__unused resultSets totalResultSetLength:(NSUInteger)__unused totalResultSetLength resultType:(ZGSearchResultType)__unused resultType dataType:(ZGVariableType)__unused dataType addressType:(ZGSearchResultAddressType)__unused addressType stride:(ZGMemorySize)__unused stride headerAddresses:(NSArray<NSNumber *> * _Nullable)__unused headerAddresses
 {
-	_progressIndicator.doubleValue = searchProgress.progress;
+	_progressIndicator.doubleValue = (double)searchProgress.progress;
 }
 
 @end
